requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
chromadb==0.4.18
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
sentence-transformers>=2.7.0
openai>=1.6.1
python-dotenv==1.0.0
httpx==0.25.2
atlassian-python-api==3.41.0
PyGithub==1.59.1
pandas==2.1.4
numpy==1.24.3
tiktoken==0.5.2
python-multipart==0.0.6
Jinja2==3.1.2

